Support executables that start with a number
The regex in installers.rs that validated the name of the executable
wouldn't allow names that started with a number. This is a problem for
executables such as `3dupdate`. There doesn't appear to be a good reason
for this restriction, so this change-set removes it.

Other restrictions, such as executables that start with an uppercase
letter, or any other non-alphanumeric character, are still in place.

This change-set also adds a few regression tests for the
`validate_name` function.